Merged PR 33748: Feat/admin-ecosystems-sidebar
:warning: **Please, review this PR after [this one](https://dev.azure.com/dcslsoftwareltd/DCSL%20Guidesmiths%20Skills%20Matrix/_git/gs-skill-matrix-frontend/pullrequest/33736) is merged. This branch is a child of that one**

This component is almost identical to the one for the regular user page adding the plus icon to add a new ecosystem. We should reuse that component when the @<Sofia Sánchez Urbano>'s PR for the user page is merged. In the meanwhile, I just added it as a non generic component to avoid too many conflict when merging the other PR.

This PR includes:

- Add ecosystems slice and store with the get request to fetch all the ecosystems
- Add the middleware endpoint and config
- Refactor the Icon component to allow reusing (including the `className` parameter)
- Refactor the Admin page styles (and SuggestionsInbox) to use CSS grid for a better layout
- Add the EcosystemsSideBar component:
  - Although it includes the plus icon, the functionality is missing. I need to check with @<Raquel> the behaviour when adding a new ecosystem.

Related work items: #109115